Search hover styling and fixing delete record bug on new note route

+++ packages/notes-frontend/app/routes/notes/new.js
@@ -18,7 +18,7 @@ export default class NotesNewRoute extends Route {
     @action
     willTransition() {
         let note = this.controller.model.note;
-        if (note && !note.content && !note.tags.length) {
+        if (note && !note.content && !note.tags.length && !note.isDeleted) {
             this.controller.model.note.destroyRecord();
         } else {
             this.controller.saveNoteTask.perform(note);
